Lab 04 - req 07 historial pedido y ver detalle andan cuando se le canta (terminado en teoria)

+++ app/src/main/java/ar/edu/utn/frsf/dam/isi/laboratorio02/dao/PedidoDAO.java
@@ -17,7 +17,7 @@ public interface PedidoDAO {
     List<Pedido> getAll();
 
     @Query("SELECT * FROM Pedido WHERE ID_PEDIDO = :PedID")
-    Pedido buscarPorID(Integer PedID);
+    Pedido buscarPorID(long PedID);
 
     @Query("SELECT * FROM Pedido WHERE ID_PEDIDO = :PedID")
     PedidoConDetalles buscarPorIDConDetalle(Integer PedID);
